% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfrj.R
\docType{package}
\name{mrfrj-package}
\alias{mrfrj}
\alias{mrfrj-package}
\title{mrfrj: Reversible Jump Monte Carlo Markov Chain for 2-dimensional Markov Random Field Interaction Structure Selection}
\description{
Implements the Reversible Jump algorithm based on pseudoposteriors
from \insertCite{paper}{mrfrj},
used to generate a Markov Chain of subsets of parameters for
sparse relative position sets.
}
\references{
\insertAllCited{}
}
\keyword{internal}
