% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mrfrj}
\alias{plot.mrfrj}
\title{Plotting `mrfrj` Objects.}
\usage{
\method{plot}{mrfrj}(x, thin = 5, ...)
}
\arguments{
\item{x}{A `mrfrj` object.}

\item{thin}{An integer representing the interval between observations in the
chain to be used. If `thin = 1`, all observations are used.}

\item{...}{Unused.}
}
\description{
Plots the RJMCMC chain produced by the `mrfrj()` function.
}
\details{
`ggplot2` is used to produce the plot, thus the method returns
a `ggplot` object which can be further modified with standard `ggplot2`
layers.
}
\author{
Victor Freguglia
}
