% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mrfi}
\alias{expand_array}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{mrf2d}{\code{\link[mrf2d:smr_array]{expand_array}}, \code{\link[mrf2d:mrfi-class]{mrfi}}}
}}

