% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rpos_summary}
\alias{rpos_summary}
\alias{rpos_plot}
\title{Relative Positions Summary and Plot}
\usage{
rpos_summary(chain, thin = 5, start_at = 0)

rpos_plot(chain, thin = 5, start_at = 0)
}
\arguments{
\item{chain}{A `\code{mrfrj}` object.}

\item{thin}{An integer representing the interval between observations in the
chain to be used. If `thin = 1`, all observations are used.}

\item{start_at}{First observation in the chain to be considered in the
subsample.}
}
\description{
Computes the marginal proportions (estimates of the probability)
of each relative position in the pseudoposterior sample generated.
The results are returned in table format for \code{rpos_summary} and
as a plot for \code{rpos_plot}.
}
\details{
To speed up computations and reduce serial correlations, use
the \code{thin} and \code{start_at} parameters to obtain proportions
based on a subsample of the chain.
}
