% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texture2d.R
\docType{data}
\name{texture2d}
\alias{texture2d}
\title{Example object for `mrfrj`}
\description{
A texture image that can be used as an example to test
the reversible jump algorithm. The data was obtained by running
the Hidden MRF algorithm from \insertCite{freguglia2020hidden}{mrfrj} with the data
from the original paper.
}
\examples{
mrf2d::dplot(texture2d)

}
\references{
\insertAllCited{}
}
\author{
Victor Freguglia
}
